% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r4dstutoriales-package.R
\docType{package}
\name{r4dstutoriales-package}
\alias{r4dstutoriales}
\alias{r4dstutoriales-package}
\title{r4dstutoriales: Tutoriales para R4DS primera edición}
\description{
Tutoriales que acompañan a la versión en español de la primera edición del libro R para Ciencia de Datos.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/cienciadedatos/r4dstutoriales}
  \item Report bugs at \url{https://github.com/cienciadedatos/r4dstutoriales/issues}
}

}
\author{
\strong{Maintainer}: Xavier de Pedro Puente \email{xavier.depedro@seeds4c.org} (\href{https://orcid.org/0009-0005-3406-2570}{ORCID})

}
\keyword{internal}
